+++ torchtune/utils/__init__.py
@@ -37,7 +37,7 @@ from .distributed import (  # noqa
     wrap_fsdp,
 )
 from .logging import get_logger
-from .memory import set_activation_checkpointing
+from .memory import set_activation_checkpointing, memory_stats_log
 from .precision import (
     get_autocast,
     get_dtype,
